Rename client to dm, update README.md and fix more minor issues.
