+++ Makefile
@@ -4,9 +4,13 @@ TARGETS := all build test push container clean
 SUBDIRS_TARGETS := \
 	$(foreach t,$(TARGETS),$(addsuffix $t,$(SUBDIRS)))
 
-GO_DEPS := github.com/kubernetes/deployment-manager/util/... github.com/kubernetes/deployment-manager/version/... github.com/kubernetes/deployment-manager/expandybird/... github.com/kubernetes/deployment-manager/resourcifier/... github.com/kubernetes/deployment-manager/manager/... github.com/kubernetes/deployment-manager/client/...
+GO_DEPS := github.com/kubernetes/deployment-manager/util/... github.com/kubernetes/deployment-manager/version/... github.com/kubernetes/deployment-manager/expandybird/... github.com/kubernetes/deployment-manager/resourcifier/... github.com/kubernetes/deployment-manager/manager/... github.com/kubernetes/deployment-manager/dm/...
 
 .PHONY : all build test clean $(TARGETS) $(SUBDIRS_TARGETS) .project .docker
+
+build:
+	go get -v $(GO_DEPS)
+	go install -v $(GO_DEPS)
 
 all: build
 
@@ -15,10 +19,6 @@ clean:
 
 test: build
 	-go test -v $(GO_DEPS)
-
-build:
-	go get -v $(GO_DEPS)
-	go install -v $(GO_DEPS)
 
 push: container
