+++ registry/github_registry.go
