+++ registry/registry.go
